full/requirements.txt
colorlog~=6.6.0
pytest~=7.4.2
PyYAML~=6.0
pylint~=2.17.4
pytest-rerunfailures~=12.0
requests~=2.27.1
allure-pytest~=2.13.2
pytest-html